feat: added javadoc comments, package structure and class-decorator
