--- src/main/java/ConsolePrinter.java
@@ -1,5 +1,0 @@
-public class ConsolePrinter {
-    public static void print(String message) {
-        System.out.println(message);
-    }
-}
