+++ src/main/java/com/tcs/edu/printer/ConsolePrinter.java
@@ -1,0 +1,16 @@
+package com.tcs.edu.printer;
+
+/**
+ * Class contains method for printing message to console
+ * @author Artashevich Viktor
+ */
+public class ConsolePrinter {
+    /**
+     * Prints the message
+     * No side-effects
+     * @param message the message to be printed
+     */
+    public static void print(String message) {
+        System.out.println(message);
+    }
+}
